backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1

pymongo==4.8.0
PyJWT==2.9.0
werkzeug==3.0.3

pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.1
joblib==1.4.2
rapidfuzz==3.9.6

PyPDF2==3.0.1
